+++ commands/create/templates/sn/picasso/common/test/integration/setup.int.js
@@ -1,3 +1,4 @@
+const path = require('path');
 const serve = require('@nebula.js/cli-serve'); // eslint-disable-line
 
 page.on('pageerror', (e) => {
@@ -17,6 +18,8 @@ if (!process.env.BASE_URL) {
     this.timeout(15000); // to allow time for the server to start
     s = await serve({
       build: false,
+      open: false,
+      fixturePath: path.resolve(__dirname, 'fixtures'),
     });
 
     process.env.BASE_URL = s.url;