+++ commands/serve/lib/init-config.js
@@ -62,9 +62,10 @@ const options = {
   webIntegrationId: {
     type: 'string',
   },
-  ACCEPT_EULA: {
-    type: 'boolean',
-    default: false,
+  fixturePath: {
+    type: 'string',
+    default: 'test/component',
+    description: 'Path to a folder that will be used as basis when locating fixtures',
   },
 };