+++ scripts/start-mashup.js
@@ -3,27 +3,15 @@
 const yargs = require('yargs');
 
 const mashupServer = require('../test/mashup/server');
-const sseServer = require('../test/fixtures/sse');
-const useEngine = require('../commands/serve/lib/engine');
 
 const args = yargs
   .option('start', {
     default: true,
     type: 'boolean',
     describe: 'Start the mashup server',
-  })
-  .option('sse', {
-    default: true,
-    type: 'boolean',
-    describe: 'Start the SSE plugin',
-  })
-  .option('engine', {
-    default: false,
-    type: 'boolean',
-    describe: 'Start the engine',
   }).argv;
 
-const { start, sse, engine } = args;
+const { start } = args;
 
 let hasCleanedUp = false;
 const services = [];
@@ -38,34 +26,6 @@ if (start) {
     },
     stop: async () => {
       await m.close();
-    },
-  });
-}
-
-if (sse) {
-  const s = sseServer;
-  services.push({
-    name: 'SSE plugin server',
-    start: () => {
-      s.start();
-    },
-    stop: () => s.close(),
-  });
-}
-
-if (engine) {
-  let e;
-  services.push({
-    name: 'Qlik Engine',
-    start: async () => {
-      e = await useEngine({
-        ACCEPT_EULA: true,
-        files: ['./test/fixtures/docker/docker-compose.yml'],
-        cwd: process.cwd(),
-      });
-    },
-    stop: async () => {
-      e && (await e());
     },
   });
 }