+++ test/component/object/sn-error.fix.js
@@ -1,8 +1,9 @@
 export default function fixture() {
   return {
     type: 'error-sn',
-    sn() {
-      throw new Error('hahaha');
-    },
+    load: async () =>
+      function sn() {
+        throw new Error('hahaha');
+      },
   };
 }