+++ apis/enigma-mocker/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/qlik-oss/nebula.js/compare/v2.2.0...v2.3.0) (2021-12-06)
+
+### Bug Fixes
+
+- **enigma-mocker:** remove build watch script ([#711](https://github.com/qlik-oss/nebula.js/issues/711)) ([8e2a3c9](https://github.com/qlik-oss/nebula.js/commit/8e2a3c962133db4d7c0e34071a313f36a25506d7))
+
 # [2.2.0](https://github.com/qlik-oss/nebula.js/compare/v2.1.0...v2.2.0) (2021-11-17)
 
 ### Features